# Conflicts:
#	code/client/api.c
#	code/client/network.c
#	code/client/player.h
#	code/common/log.c
#	code/common/uuid.h
#	include/client/event.h

+++ include/client/Headquarter.h
@@ -19,10 +19,12 @@
 #  define __cdecl __attribute__((__cdecl__))
 # endif
 
+#include <stdint.h>
+
 #include "str.h"
 #include <common/array.h>
 #include <common/assert.h>
-#include <common/dlfunc.h>
+#include <common/paths.h>
 #include <common/noreturn.h>
 
 #ifndef _Noreturn
@@ -133,7 +135,7 @@ HQAPI uint32_t          GetNpcIdOfAgent(AgentId agent_id);
 HQAPI bool              GetItem(ApiItem *item, uint32_t item_id);
 HQAPI bool              GetItemOfAgent(ApiItem *item, AgentId agent_id);
 HQAPI BagEnum           GetItemLocation(uint32_t item_id, unsigned int *slot);
-HQAPI int            GetItemModStruct(uint32_t item_id, uint32_t *buffer, size_t length);
+HQAPI size_t            GetItemModStruct(uint32_t item_id, uint32_t *buffer, size_t length);
 HQAPI size_t            GetItemName(uint32_t item_id, uint16_t* buffer, size_t length);
 HQAPI size_t            GetBagCapacity(BagEnum bag);
 HQAPI size_t            GetBagItems(BagEnum bag, ApiItem *buffer, size_t length);
